// client/src/zod/registerSchema.ts
import {z} from "zod"

export const registerSchema = z.object({
    username: z.string().min(3, "Username must be at least 3 characters!"),
    email: z.string().email("Invalid email address"),
    password: z.string().min(6, "Password must be at least 6 characters!"),
    confirmPassword: z.string()
})
. refine( (data) => data.password === data.confirmPassword, {
    message: "Passwords do not match",
    path: ["confirmPassword"], // show error under confirmPassword field
})

export type RegisterFormData = z.infer<typeof registerSchema>